+++ mithril-stm/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.5.3 (10-07-2025)
+
+### Added
+
+- Added support for multiple aggregate signature proof systems.
+
 ## 0.5.0 (09-09-2025)
 
 ### Changed